# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jev Triage API",
    "description": "Tickets on a Kanban board, triaged by TypeSafe AI's Jev.",
    "version": "0.1.0"
  },
  "paths": {
    "/meta": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Read Meta",
        "operationId": "read_meta",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetaRead"
                }
              }
            }
          }
        }
      }
    },
    "/tickets": {
      "get": {
        "tags": [
          "tickets"
        ],
        "summary": "List Tickets",
        "operationId": "list_tickets",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TicketRead"
                  },
                  "type": "array",
                  "title": "Response List Tickets"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tickets"
        ],
        "summary": "Create Ticket",
        "operationId": "create_ticket",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickets/{ticket_id}": {
      "get": {
        "tags": [
          "tickets"
        ],
        "summary": "Read Ticket",
        "operationId": "read_ticket",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "tickets"
        ],
        "summary": "Update Ticket",
        "operationId": "update_ticket",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tickets"
        ],
        "summary": "Delete Ticket",
        "operationId": "delete_ticket",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickets/{ticket_id}/move": {
      "post": {
        "tags": [
          "tickets"
        ],
        "summary": "Move Ticket",
        "description": "Moving into Triaged runs triage once. Any other move is free. Failure leaves the ticket.",
        "operationId": "move_ticket",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickets/{ticket_id}/triage": {
      "post": {
        "tags": [
          "tickets"
        ],
        "summary": "Retriage Ticket",
        "description": "Always asks the provider again and replaces the stored result.",
        "operationId": "retriage_ticket",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MetaRead": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "mock": {
            "type": "boolean",
            "title": "Mock"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "mock"
        ],
        "title": "MetaRead"
      },
      "MoveRequest": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "MoveRequest"
      },
      "TicketCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description"
        ],
        "title": "TicketCreate"
      },
      "TicketRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "department_override": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "billing",
                  "technical",
                  "general"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Override"
          },
          "triage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TriageResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "effective_department": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "billing",
                  "technical",
                  "general"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Department"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "status",
          "department_override",
          "triage",
          "effective_department",
          "created_at",
          "updated_at"
        ],
        "title": "TicketRead"
      },
      "TicketStatus": {
        "type": "string",
        "enum": [
          "new",
          "triaged",
          "in_progress",
          "done"
        ],
        "title": "TicketStatus"
      },
      "TicketUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "department_override": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "billing",
                  "technical",
                  "general"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Override"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TicketUpdate",
        "description": "PATCH body. Fields left out are untouched; department_override may be set to null."
      },
      "TriageResult": {
        "properties": {
          "department": {
            "type": "string",
            "title": "Department"
          },
          "department_probabilities": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Department Probabilities"
          },
          "urgency_score": {
            "type": "number",
            "title": "Urgency Score"
          },
          "urgency_level": {
            "type": "integer",
            "title": "Urgency Level"
          },
          "urgency_mean": {
            "type": "number",
            "title": "Urgency Mean"
          },
          "urgency_probabilities": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Urgency Probabilities"
          },
          "refund_requested": {
            "type": "boolean",
            "title": "Refund Requested"
          },
          "refund_probability": {
            "type": "number",
            "title": "Refund Probability"
          },
          "confidence": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "triaged_at": {
            "type": "string",
            "format": "date-time",
            "title": "Triaged At"
          }
        },
        "type": "object",
        "required": [
          "department",
          "department_probabilities",
          "urgency_score",
          "urgency_level",
          "urgency_mean",
          "urgency_probabilities",
          "refund_requested",
          "refund_probability",
          "provider",
          "triaged_at"
        ],
        "title": "TriageResult"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}